major architecture rework: add Input namespace with one struct that handles keyboard event data. Delete view/monitor folder, now renderer and static-data locates directly in src.
