+++ src/input/input.h
@@ -1,0 +1,9 @@
+#pragma once
+#include <iostream>
+
+namespace Input {
+    struct KeyboardEvent {
+        uint8_t code;
+        uint8_t type;
+    };
+}
