+++ src/renderer/debug-renderer/debug-renderer.h
@@ -1,0 +1,8 @@
+#pragma once
+#include "renderer/irenderer/irenderer.h"
+
+namespace Renderer {
+    class DebugRenderer : IRenderer {
+
+    };
+}
