+++ src/renderer/irenderer/irenderer.h
@@ -1,13 +1,10 @@
 #pragma once
 #include <iostream>
-#include <SFML/Graphics.hpp>
+#include "SFML/Graphics.hpp"
+#include "input/input.h"
 
-struct KeyboardEvent {
-    uint8_t type;
-    uint8_t code;
-};
-
-class IRenderer {
+namespace Renderer {
+    class IRenderer {
     public:
         // constructors
         IRenderer();
@@ -16,7 +13,7 @@ class IRenderer {
         virtual void start();
 
         // call in main event loop. Updates the screen. IDK is it good idea to implement it with double, but why not
-        virtual KeyboardEvent update(uint64_t deltatime) = 0;
+        virtual Input::KeyboardEvent update(uint64_t deltatime) = 0;
 
         // close window at program finish
         virtual void close();
@@ -25,16 +22,18 @@ class IRenderer {
         virtual int drawSprite(int x, int y, std::string sprite_id) = 0;
 
         // call to draw rectangle
-        virtual int drawRec(int x, int y, int w, int h, std::string color="#000000", std::string stroke_color="NONE", int stroke_weight=0) = 0;
+        virtual int
+        drawRec(int x, int y, int w, int h, std::string color = "#000000", std::string stroke_color = "NONE",
+                int stroke_weight = 0) = 0;
 
         // call to draw
         virtual int drawText(int x, int y, std::string line) = 0;
 
         // check if is active
-        virtual bool isActive ();
+        virtual bool isActive();
 
         // get previous keyboard event
-        KeyboardEvent getPrevKeyboardEvent() ;
+        Input::KeyboardEvent getPrevKeyboardEvent();
 
         // destructor
         virtual ~IRenderer() = default;
@@ -47,5 +46,6 @@ class IRenderer {
         bool active;
 
         // previous keyboard event
-        KeyboardEvent prev_keyboard_event;
-};
+        Input::KeyboardEvent prev_keyboard_event;
+    };
+}
