+++ src/renderer/renderer.h
@@ -1,0 +1,4 @@
+#pragma once
+#include "sfml-renderer/sfml-renderer.h"
+#include "irenderer/irenderer.h"
+#include "debug-renderer/debug-renderer.h"
