--- src/view/renderer/debug-renderer/debug-renderer.h
@@ -1,6 +1,0 @@
-#pragma once
-#include "view/renderer/irenderer/irenderer.h"
-
-class DebugRenderer : IRenderer {
-
-};
